changed color-channels to byte

+++ Extended/Graphics/UI/UIDim.cs
@@ -4,7 +4,7 @@ using mapKnight.Extended.Graphics.UI.Layout;
 
 namespace mapKnight.Extended.Graphics.UI {
     public class UIDim : UIItem {
-        public float Opacity { get { return color.A / 255f; } set { color.A = (int)((1f-value) * 255); IsDirty = true; } }
+        public float Opacity { get { return color.A / 255f; } set { color.A = (byte)((1f-value) * 255); IsDirty = true; } }
         private Color color;
 
         public UIDim (Screen owner, float opacity, int depth, bool multiclick = false) : base(owner, new UILeftMargin(0), new UITopMargin(0), new Vector2(5, 5), depth, multiclick) {
